Refactor: Remove redundant react-usememo-should-be-component-bug test file
The test cases previously in the separate `react-usememo-should-be-component-bug.test.ts` file have been merged into the main `react-usememo-should-be-component.test.ts` test suite. This eliminates redundancy and simplifies the test file structure.
